Fix CR suggestion

+++ src/main/java/com/salesforce/dataloader/security/EncryptionAesUtil.java
@@ -32,19 +32,13 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
+import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.security.GeneralSecurityException;
-import java.security.InvalidAlgorithmParameterException;
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
-import java.security.spec.InvalidKeySpecException;
-import java.security.spec.InvalidParameterSpecException;
+import java.util.Locale;
 
-import javax.crypto.BadPaddingException;
 import javax.crypto.Cipher;
-import javax.crypto.IllegalBlockSizeException;
-import javax.crypto.NoSuchPaddingException;
 import javax.crypto.spec.IvParameterSpec;
 import javax.crypto.spec.SecretKeySpec;
 
@@ -57,22 +51,51 @@ public class EncryptionAesUtil {
 
     private static final Logger LOGGER = Logger.getLogger(EncryptionAesUtil.class);
 
+    public enum OSType {
+        Windows, MacOS, Linux, Other
+    }
+
+    ;
+    protected static OSType detectedOS;
+
+    /**
+     * detect the operating system from the os.name System property and cache the result
+     *
+     * @returns - the operating system detected
+     */
+    public static OSType getOperatingSystemType() {
+        if (detectedOS == null) {
+            String OS = System.getProperty("os.name", "generic").toLowerCase(Locale.ENGLISH);
+            if ((OS.indexOf("mac") >= 0) || (OS.indexOf("darwin") >= 0)) {
+                detectedOS = OSType.MacOS;
+            } else if (OS.indexOf("win") >= 0) {
+                detectedOS = OSType.Windows;
+            } else if (OS.indexOf("nux") >= 0) {
+                detectedOS = OSType.Linux;
+            } else {
+                detectedOS = OSType.Other;
+            }
+        }
+        return detectedOS;
+    }
+
     // Support single text encryption and decryption
 
     static public Cipher cipher;
     static public byte[] cipherKey;
 
-
+    // 16 bytes was used for 128 bit AES encryption
     public static final int ENCRYPTION_KEY_LENGTH_IN_BYTES = 16;
     public static final String DEFAULT_KEYFILE_NAME = "dataLoader.key";
 
     static {
         try {
             cipher = Cipher.getInstance("AES/CBC/PKCS5Padding");
+            detectedOS = getOperatingSystemType();
 
         } catch (Exception e) {
             LOGGER.error("Fail to initialize encryption: " + e.getMessage());
-            throw new  RuntimeException("Fail to initialize encryption:  ", e);
+            throw new RuntimeException("Fail to initialize encryption:  ", e);
         }
     }
 
@@ -100,6 +123,9 @@ public class EncryptionAesUtil {
 
     public byte[] generateEncryptionKey() {
         // The secure random uses threadId and tick to make sure it is not repeatable
+        // If a call to setSeed had not occurred previously,
+        // the first call to this method forces this SecureRandom object to seed itself.
+        // This self-seeding will not occur if setSeed was previously called.
         SecureRandom random = new SecureRandom();
         byte bytes[] = new byte[ENCRYPTION_KEY_LENGTH_IN_BYTES];
         random.nextBytes(bytes);
@@ -107,13 +133,20 @@ public class EncryptionAesUtil {
     }
 
     public String createUserProfileKeyName() {
-        String path = Paths.get(System.getProperty("user.home"), "dataloader").toString();
+        String path = Paths.get(System.getProperty("user.home"), ".dataloader").toString();
         File customDir = new File(path);
 
         if (customDir.exists()) {
-            LOGGER.info(customDir + " exists");
+            LOGGER.debug(customDir + " exists");
         } else if (customDir.mkdirs()) {
+            if (detectedOS == OSType.MacOS || detectedOS == OSType.Linux) {
+                // set all reading to false
+                customDir.setReadable(false, false);
+                // only owner can read
+                customDir.setReadable(true, true);
+            }
             LOGGER.info(customDir + " was created");
+
         } else {
             LOGGER.info(customDir + " was not created");
             throw new RuntimeException("Cannot create directory:" + path);
@@ -121,47 +154,75 @@ public class EncryptionAesUtil {
         return Paths.get(path, DEFAULT_KEYFILE_NAME).toString();
     }
 
-    public String createKeyFile(String filePath) throws IOException {
-        if (filePath == null || filePath.isEmpty()) {
-            filePath = createUserProfileKeyName();
+    // If given path not existing, create one.
+    public String createKeyFileIfNotExisting(String filePath) throws GeneralSecurityException {
+        if (filePath == null || filePath.isEmpty() || !Files.exists(Paths.get(filePath))) {
+            // no valid file path provided, check if default one exists
+            if (filePath == null || filePath.isEmpty()) {
+                filePath = createUserProfileKeyName();
+            }
+            if (!Files.exists(Paths.get(filePath))) {
+                byte[] key = generateEncryptionKey();
+                try (FileOutputStream fos = new FileOutputStream(filePath)) {
+                    fos.write(key);
+                } catch (IOException io) {
+                    throw new GeneralSecurityException("Failed to open file:" + filePath, io);
+                }
+                // Windows platform is already readable only to owner.
+                if (detectedOS == OSType.MacOS || detectedOS == OSType.Linux) {
+                    File file = new File(filePath);
+                    // set all reading to false
+                    file.setReadable(false, false);
+                    // set Owner reading to true
+                    file.setReadable(true, true);
+                }
+            }
         }
-        byte[] key = generateEncryptionKey();
-        try (FileOutputStream fos = new FileOutputStream(filePath)) {
-            fos.write(key);
-        }
+        setCipherKeyFromFilePath(filePath);
         return filePath;
     }
 
-    public void resetCipherKey(){
+    public void resetCipherKey() {
         cipherKey = null;
     }
 
-    public String setCipherKeyFromFilePath(String filePath) throws IOException, GeneralSecurityException {
+    public void setCipherKeyFromFilePath(String filePath) throws GeneralSecurityException {
         byte[] data = new byte[1024];
         try (FileInputStream fis = new FileInputStream(filePath)) {
             int size = fis.read(data);
             if (size < ENCRYPTION_KEY_LENGTH_IN_BYTES)
                 throw new GeneralSecurityException("Keyfile content is too short:" + filePath);
-            else
-            {
+            else {
                 cipherKey = new byte[ENCRYPTION_KEY_LENGTH_IN_BYTES];
                 System.arraycopy(data, 0, cipherKey, 0, ENCRYPTION_KEY_LENGTH_IN_BYTES);
             }
-            return new String(cipherKey, "UTF-8");
+        } catch (IOException io) {
+            throw new GeneralSecurityException("Failed to open file: " + filePath, io);
         }
+    }
+
+    private void ensureKeyIsSet() throws IOException, GeneralSecurityException {
+        if (cipherKey != null) {
+            return;
+        }
+        // if no key was set up already, create one
+        createKeyFileIfNotExisting(null);
+
     }
 
     public String encryptMsg(String msg) throws GeneralSecurityException {
         try {
+            ensureKeyIsSet();
             return EncryptionUtil.bytesToText(encryptMsg(msg, cipherKey));
         } catch (Exception e) {
             LOGGER.error("Fail to encrypt message: " + e.getMessage());
-            throw new  GeneralSecurityException("Error to encrypt message: ", e);
+            throw new GeneralSecurityException("Error to encrypt message: ", e);
         }
     }
 
     public String decryptMsg(String cipherMsgString) throws GeneralSecurityException {
         try {
+            ensureKeyIsSet();
             byte[] cipherMsg = EncryptionUtil.textToBytes(cipherMsgString);
             return decryptMsg(cipherMsg, cipherKey);
         } catch (Exception e) {
@@ -170,26 +231,25 @@ public class EncryptionAesUtil {
         }
     }
 
-    synchronized public byte[] encryptMsg(String msg, byte[] encryptionKey) throws NoSuchPaddingException, NoSuchAlgorithmException, InvalidKeySpecException, InvalidKeyException, InvalidParameterSpecException, UnsupportedEncodingException, BadPaddingException, IllegalBlockSizeException {
+    synchronized public byte[] encryptMsg(String msg, byte[] encryptionKey) throws GeneralSecurityException, UnsupportedEncodingException {
 
-        if(encryptionKey == null)
-            return msg.getBytes("UTF-8");
+        if (encryptionKey == null || encryptionKey.length != ENCRYPTION_KEY_LENGTH_IN_BYTES)
+            throw new GeneralSecurityException("Encryption key is null or has invalid length");
         SecretKeySpec key = new SecretKeySpec(encryptionKey, "AES");
-        cipher.init(Cipher.ENCRYPT_MODE, key);
-
-        byte[] cipherText = cipher.doFinal(msg.getBytes("UTF-8"));
-        byte[] iv = cipher.getIV();
-        return concatenateByteArray(iv, cipherText);
+        byte[] ivBytes = generateEncryptionKey();
+        IvParameterSpec iv = new IvParameterSpec(ivBytes);
+        cipher.init(Cipher.ENCRYPT_MODE, key, iv);
+        byte[] cipherText = cipher.doFinal(msg.getBytes());
+        return concatenateByteArray(ivBytes, cipherText);
     }
 
-    synchronized public String decryptMsg(byte[] cipherMsg, byte[] encryptionKey) throws InvalidKeySpecException, NoSuchAlgorithmException, BadPaddingException, IllegalBlockSizeException, InvalidAlgorithmParameterException, InvalidKeyException, UnsupportedEncodingException {
+    synchronized public String decryptMsg(byte[] cipherMsg, byte[] encryptionKey) throws GeneralSecurityException, UnsupportedEncodingException {
         if (encryptionKey == null)
-            return new String(cipherMsg,"UTF-8");
+            throw new GeneralSecurityException("Encryption key is null or has invalid length");
         SecretKeySpec key = new SecretKeySpec(encryptionKey, "AES");
         byte[] iv = extractIvBytes(cipherMsg);
         cipher.init(Cipher.DECRYPT_MODE, key, new IvParameterSpec(iv));
         byte[] cipherContent = extractCipherContent(cipherMsg);
-        String plaintext = new String(cipher.doFinal(cipherContent), "UTF-8");
-        return plaintext;
+        return new String(cipher.doFinal(cipherContent));
     }
 }
